github-integration: split discussions module
